less warnings in msvs
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@1374 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ afanasy/trunk/src/libafanasy/addresseslist.h
@@ -18,7 +18,7 @@ public:
 
    bool addAddress( const Address * address );
 
-   inline int getAddressesNum() const { return addresses.size(); }
+   inline size_t getAddressesNum() const { return addresses.size(); }
 
    inline const std::list<Address*> * getAddresses() const { return &addresses;}
